Create three different blink targets
