+++ .gitignore
@@ -1,6 +1,7 @@
 Testing
 build
 log
+.vscode
 
 *.log.*
 *-asciinema
